фикс пр
