+++ packages/client/src/pages/user/styles.ts
@@ -1,0 +1,19 @@
+const styles = {
+  grid: {
+    padding: '16px',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  container_center: {
+    borderRadius: '16px',
+    textAlign: 'center',
+  },
+  container: {
+    borderRadius: '16px',
+  },
+  grid_item: {
+    width: '50%',
+  },
+} as const
+
+export default styles
