+++ packages/client/src/pages/user/user.tsx
@@ -1,131 +1,40 @@
 import {
-  Avatar,
   Box,
-  Button,
   Card,
   CardContent,
   CardHeader,
-  CardMedia,
-  Chip,
-  Container,
   Divider,
   Grid,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  TextField,
   Typography,
 } from '@mui/material'
-import { ROWS, RECORDS } from './model/constants'
-import { Create, EmojiEvents } from '@mui/icons-material'
-import { styled } from '@mui/material/styles'
-import {
-  changePassword,
-  getUserInfo,
-  updateAvatar,
-} from '../../entities/user/api'
+import { EmojiEvents } from '@mui/icons-material'
+import { getUserInfo } from '../../entities/user/api'
 import { useEffect, useState } from 'react'
+import { IUser } from './model/userData'
+import UserAvatar from '../../widgets/avatar/avatar'
+import ChangePassword from '../../widgets/change-password/change-pasword'
+import RecordsTable from '../../widgets/records-table/records-table'
+import styles from './styles'
 
-const BASE_URL = 'https://ya-praktikum.tech/api/v2'
-type PropsType = {
-  data?: any
-}
-
-interface User {
-  avatar?: string
-  display_name?: string
-  email?: string
-  first_name?: string
-  id?: number
-  login?: string
-  phone?: string
-  second_name?: string
-}
-
-const VisuallyHiddenInput = styled('input')({
-  clip: 'rect(0 0 0 0)',
-  clipPath: 'inset(50%)',
-  height: 1,
-  overflow: 'hidden',
-  position: 'absolute',
-  bottom: 0,
-  left: 0,
-  whiteSpace: 'nowrap',
-  width: 1,
-})
-
-const User = ({ data }: PropsType) => {
-  const [oldPassword, setOldPassword] = useState('')
-  const [newPassword, setNewPassword] = useState('')
-  const [user, setUser] = useState<User>({})
-  const [isAvatarChanging, setIsAvatarChanging] = useState(false)
-  const [avatarFormData, setAvatarFormData] = useState<FormData>(new FormData())
-  const [avatar, setAvatar] = useState<string>('')
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = event.target
-    if (name === 'oldPassword') {
-      setOldPassword(value)
-    } else if (name === 'newPassword') {
-      setNewPassword(value)
-    }
-  }
-
-  const handleSubmit = (event: React.MouseEvent) => {
-    event.preventDefault()
-    changePassword({ oldPassword, newPassword })
-    setNewPassword('')
-    setOldPassword('')
-  }
-
-  const handleAvatarChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0]
-    if (file) {
-      const formData = new FormData()
-      formData.append('avatar', file)
-      setIsAvatarChanging(true)
-      setUser(prevUser => ({
-        ...prevUser,
-        avatar: URL.createObjectURL(file),
-      }))
-      setAvatarFormData(formData)
-    }
-  }
-
-  const handleUpdateAvatar = async () => {
-    const response = await updateAvatar(avatarFormData)
-    const data = await response.json()
-    setUser(prevUser => ({ ...prevUser, avatar: data.avatar }))
-    setIsAvatarChanging(false)
-  }
+const User = () => {
+  const [user, setUser] = useState<IUser>({})
 
   useEffect(() => {
     const fetchUserInfo = async () => {
-      const userInfo = await getUserInfo()
-      setUser(userInfo)
+      try {
+        const userInfo = await getUserInfo()
+        setUser(userInfo)
+      } catch (error) {
+        throw new Error(`Ошибка запроса! статус: ${error}`)
+      }
     }
     fetchUserInfo()
   }, [])
 
   return (
-    <Grid
-      container
-      spacing={4}
-      direction="column"
-      sx={{
-        padding: '16px',
-        justifyContent: 'center',
-        alignItems: 'center',
-      }}>
-      <Grid item xs={6} sx={{ width: '50%' }}>
-        <Card
-          sx={{
-            borderRadius: '16px',
-            textAlign: 'center',
-          }}>
+    <Grid container spacing={4} direction="column" sx={styles.grid}>
+      <Grid item xs={6} sx={styles.grid_item}>
+        <Card sx={styles.container_center}>
           <CardHeader
             title={
               <Typography variant="h6" component="span" color={'primary'}>
@@ -136,70 +45,12 @@ const User = ({ data }: PropsType) => {
 
           <Divider orientation="horizontal" flexItem />
 
-          {user && (
-            <CardMedia
-              sx={{
-                display: 'flex',
-                alignItems: 'center',
-                gap: '8px',
-                flexDirection: 'column',
-                padding: '24px',
-              }}>
-              <Card
-                variant="outlined"
-                sx={{
-                  borderRadius: '50%',
-                  pointerEvents: 'none',
-                }}>
-                <Avatar
-                  src={
-                    isAvatarChanging
-                      ? user.avatar
-                      : `${BASE_URL}/resources${user.avatar}`
-                  }
-                  sx={{
-                    width: 250,
-                    height: 250,
-                  }}
-                />
-              </Card>
-
-              <Typography gutterBottom variant="caption" component="span">
-                Вы можете сменить свой аватар
-              </Typography>
-              <Button
-                component="label"
-                role={undefined}
-                variant="contained"
-                tabIndex={-1}
-                endIcon={<Create />}
-                size="small">
-                Сменить аватар
-                <VisuallyHiddenInput
-                  type="file"
-                  onChange={handleAvatarChange}
-                />
-              </Button>
-              {isAvatarChanging && avatarFormData && (
-                <Button
-                  variant="contained"
-                  color="primary"
-                  onClick={() => {
-                    handleUpdateAvatar()
-                  }}>
-                  Сохранить
-                </Button>
-              )}
-            </CardMedia>
-          )}
+          {user && <UserAvatar user={user} setUser={setUser} />}
         </Card>
       </Grid>
 
-      <Grid item xs={6} sx={{ width: '50%' }}>
-        <Card
-          sx={{
-            borderRadius: '16px',
-          }}>
+      <Grid item xs={6} sx={styles.grid_item}>
+        <Card sx={styles.container}>
           <CardHeader
             title={
               <Box sx={{ display: 'flex', alignItems: 'center' }}>
@@ -214,46 +65,13 @@ const User = ({ data }: PropsType) => {
           <Divider orientation="horizontal" flexItem />
 
           <CardContent sx={{ padding: '24px' }}>
-            <TableContainer component={Box} sx={{ maxHeight: '280px' }}>
-              <Table stickyHeader>
-                <TableHead sx={{ position: 'sticky', top: 0 }}>
-                  <TableRow>
-                    {ROWS.map(row => (
-                      <TableCell align="center">
-                        <Typography variant="caption" color={'primary'}>
-                          {row}
-                        </Typography>
-                      </TableCell>
-                    ))}
-                  </TableRow>
-                </TableHead>
-
-                <TableBody sx={{ maxHeight: '200px', overflow: 'auto' }}>
-                  {RECORDS.map((record, index) => (
-                    <TableRow key={record.id}>
-                      <TableCell align="center">{index + 1}</TableCell>
-                      <TableCell align="center">{record.date}</TableCell>
-                      <TableCell align="center">{record.value}</TableCell>
-                      <TableCell align="center">
-                        <Chip
-                          label={record.status}
-                          color={record.isWinner ? 'success' : 'default'}
-                        />
-                      </TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            </TableContainer>
+            <RecordsTable />
           </CardContent>
         </Card>
       </Grid>
 
-      <Grid item xs={6} sx={{ width: '50%' }}>
-        <Card
-          sx={{
-            borderRadius: '16px',
-          }}>
+      <Grid item xs={6} sx={styles.grid_item}>
+        <Card sx={styles.container}>
           <CardHeader
             title={
               <Typography
@@ -268,49 +86,7 @@ const User = ({ data }: PropsType) => {
 
           <Divider orientation="horizontal" flexItem />
 
-          <CardContent sx={{ padding: '24px' }}>
-            <Box
-              sx={{
-                display: 'flex',
-                flexDirection: 'row',
-                gap: 4,
-                justifyContent: 'center',
-                padding: '16px 0',
-              }}>
-              <TextField
-                id="outlined-password-input"
-                label="Старый пароль"
-                type="password"
-                name="oldPassword"
-                value={oldPassword}
-                onChange={handleChange}
-                sx={{ flexGrow: 1 }}
-              />
-
-              <TextField
-                id="outlined-password-input"
-                label="Новый пароль"
-                type="password"
-                name="newPassword"
-                value={newPassword}
-                onChange={handleChange}
-                sx={{ flexGrow: 1 }}
-              />
-            </Box>
-
-            <Divider orientation="horizontal" flexItem />
-
-            <Box
-              sx={{
-                display: 'flex',
-                justifyContent: 'flex-end',
-                padding: '16px 0',
-              }}>
-              <Button variant="contained" onClick={handleSubmit}>
-                Сохранить
-              </Button>
-            </Box>
-          </CardContent>
+          <ChangePassword />
         </Card>
       </Grid>
     </Grid>
