+++ packages/client/src/shared/constants/api.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = 'https://ya-praktikum.tech/api/v2'
