+++ packages/client/src/widgets/avatar/styles.ts
@@ -1,0 +1,19 @@
+const styles = {
+  image: {
+    width: 250,
+    height: 250,
+  },
+  image_container: {
+    borderRadius: '50%',
+    pointerEvents: 'none',
+  },
+  container: {
+    display: 'flex',
+    alignItems: 'center',
+    gap: '8px',
+    flexDirection: 'column',
+    padding: '24px',
+  },
+} as const
+
+export default styles
