+++ packages/client/src/widgets/change-password/styles.ts
@@ -1,0 +1,19 @@
+const styles = {
+  container: {
+    display: 'flex',
+    flexDirection: 'row',
+    gap: 4,
+    justifyContent: 'center',
+    padding: '16px 0',
+  },
+  input: {
+    flexGrow: 1,
+  },
+  button_container: {
+    display: 'flex',
+    justifyContent: 'flex-end',
+    padding: '16px 0',
+  },
+} as const
+
+export default styles
