+++ packages/client/src/widgets/records-table/styles.ts
@@ -1,0 +1,7 @@
+const styles = {
+  table_body: { maxHeight: '200px', overflow: 'auto' },
+  table_header: { position: 'sticky', top: 0 },
+  table_container: { maxHeight: '280px' },
+} as const
+
+export default styles
